requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
msoffcrypto-tool>=5.4.0
openpyxl>=3.1.0
plotly>=5.18.0
